+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Changed
 
 - `Quickie::open_bi` now only returns a single stream ID
+- `Quickie::disconnect` is no longer `async`
 - renamed `Quickie::unicast` to `::send_msg` (to be more aligned with `::send_datagram`)
 
 ### Fixed